yash-unified-mdoc-user-story-backend-dev/requirements-freeze.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiosignal==1.4.0
altair==5.5.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.11.0
attrs==25.4.0
blinker==1.9.0
boto3==1.41.2
botocore==1.41.2
cachetools==6.2.2
certifi==2025.11.12
charset-normalizer==3.4.4
click==8.3.1
defusedxml==0.7.1
distro==1.9.0
dotenv==0.9.9
fastapi==0.121.3
fastuuid==0.14.0
ffmpeg-python==0.2.0
filelock==3.20.0
fonttools==4.60.1
fpdf2==2.8.5
frozenlist==1.8.0
fsspec==2025.10.0
future==1.0.0
gitdb==4.0.12
GitPython==3.1.45
graphviz==0.21
h11==0.16.0
hf-xet==1.2.0
httpcore==1.0.9
httpx==0.28.1
huggingface_hub==1.1.5
idna==3.11
ImageIO==2.37.2
importlib_metadata==8.7.0
Jinja2==3.1.6
jiter==0.12.0
jmespath==1.0.1
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
lazy_loader==0.4
litellm==1.80.5
lxml==6.0.2
MarkupSafe==3.0.3
multidict==6.7.0
narwhals==2.12.0
networkx==3.5
numpy==2.2.6
openai==2.8.1
opencv-python-headless==4.12.0.88
packaging==25.0
pandas==2.3.3
pillow==12.0.0
propcache==0.4.1
protobuf==6.33.1
psutil==7.1.3
psycopg2-binary==2.9.11
pyarrow==21.0.0
pydantic==2.12.4
pydantic_core==2.41.5
pydeck==0.9.1
pydub==0.25.1
PyJWT==2.10.1
pytesseract==0.3.13
python-dateutil==2.9.0.post0
python-docx==1.2.0
python-dotenv==1.2.1
python-multipart==0.0.20
pytz==2025.2
PyYAML==6.0.3
referencing==0.37.0
regex==2025.11.3
reportlab==4.4.5
requests==2.32.5
rpds-py==0.29.0
s3transfer==0.15.0
scikit-image==0.25.2
scipy==1.16.3
shellingham==1.5.4
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
SpeechRecognition==3.14.4
starlette==0.50.0
streamlit==1.51.0
tenacity==9.1.2
tifffile==2025.10.16
tiktoken==0.12.0
tokenizers==0.22.1
toml==0.10.2
tornado==6.5.2
tqdm==4.67.1
typer-slim==0.20.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
uvicorn==0.38.0
watchdog==6.0.0
whisper==1.1.10
yarl==1.22.0
zipp==3.23.0
